backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pdfplumber
python-docx
rapidfuzz
requests
python-dotenv
